rename device.rs to devices.rs
