+++ crates/cinemotion-core/src/devices.rs
