+++ crates/cinemotion-core/src/scene_test.rs
@@ -86,7 +86,7 @@ async fn test_scene_system_attribute_links() {
     device.insert_attribute(Attribute::new_matrix44("transform"));
 
     // TODO: Rename device to devices
-    let device_id = crate::device::system::add_device(&mut world, device.clone());
+    let device_id = crate::devices::system::add_device(&mut world, device.clone());
 
     let mut object = SceneObject::new("camera1");
     object.insert_attribute(Attribute::new_matrix44("transform"));
@@ -98,7 +98,7 @@ async fn test_scene_system_attribute_links() {
     system::update(&mut world).expect("should not fail on first iteration");
     let object_ref = system::get_by_id(&mut world, id).expect("object should exist");
     let device_ref =
-        device::system::get_by_id(&mut world, &device_id).expect("device should exist");
+        devices::system::get_by_id(&mut world, &device_id).expect("device should exist");
 
     // The object's linked attribute should be updated to match the device's attribute.
     assert_eq!(
@@ -111,11 +111,11 @@ async fn test_scene_system_attribute_links() {
     let attribute = Attribute::new("transform", value);
     device.insert_attribute(attribute);
 
-    crate::device::system::set_device(&mut world, device_id.clone(), device);
+    crate::devices::system::set_device(&mut world, device_id.clone(), device);
 
     let object_ref = system::get_by_id(&mut world, id).expect("object should exist");
     let device_ref =
-        device::system::get_by_id(&mut world, &device_id).expect("device should exist");
+        devices::system::get_by_id(&mut world, &device_id).expect("device should exist");
 
     // The object's linked attribute should NOT be updated to match the device's attribute.
     assert_ne!(
@@ -129,7 +129,7 @@ async fn test_scene_system_attribute_links() {
     // The object's linked attribute should be updated to match the device's attribute.
     let object_ref = system::get_by_id(&mut world, id).expect("object should exist");
     let device_ref =
-        device::system::get_by_id(&mut world, &device_id).expect("device should exist");
+        devices::system::get_by_id(&mut world, &device_id).expect("device should exist");
     assert_eq!(
         object_ref.attribute(&world, "transform").unwrap().value(),
         device_ref.attribute(&name!("transform")).unwrap().value()
